chore(testing): split custom matchers out of setupEnv
Move the AWS SDK, Logger and Metrics matchers into one module each
under packages/testing/src/matchers, keeping each family's
implementation next to its assertion typings. setupEnv keeps its path
and only registers the merged object.
